refactor(theme): consolidate theme tests into a single comprehensive suite
